Add GroundFlame component for collider-based burn damage
