+++ Assets/Scripts/NPC/Combat/Projectiles/GroundFlame.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8876728a984446078d2af97dc69c5831
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
